RavenDB-15742 : add more tests
